backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiofiles>=24.1.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
